wip(#1101): extract lower/pre_scan
